+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   No applied tag-filter will still export all statements unchanged
   
 - Add similar submitters to the Submitter List
+- Add anonymous voters column to statement XLSX export
 - Add segment text to the boilerplate modal, if the segment and its text is available
 
 ## v4.25.0 (2025-11-06)